@@ -37,7 +37,7 @@ class RequestBean {
       method: json['method'],
       contentType: json['contentType'],
       followRedirects: json['followRedirects'],
-      requestTime: DateTime.tryParse(json['requestTime'] ?? ''),
+      requestTime: DateTime.tryParse(json['requestTime'] ?? '')?.toLocal(),
       headers: json['headers'],
       params: json['params'],
       body: formData ?? json['body'],
@@ -83,7 +83,7 @@ class RequestBean {
       'method': method,
       'contentType': contentType,
       'followRedirects': followRedirects,
-      'requestTime': requestTime?.toString(),
+      'requestTime': requestTime?.toUtc().toString(),
       'headers': changedHeaders,
       'params': changedParams,
       'body': body is FormData ? getFormData() : changedBody ?? body,
