@@ -22,7 +22,7 @@ class ResponseBean {
       url: json['url'],
       method: json['method'],
       statusMessage: json['statusMessage'],
-      responseTime: DateTime.tryParse(json['responseTime'] ?? ''),
+      responseTime: DateTime.tryParse(json['responseTime'] ?? '')?.toLocal(),
       duration: json['duration'],
       data: json['data'],
       headers: json['headers'],
@@ -58,7 +58,7 @@ class ResponseBean {
       'url': url,
       'method': method,
       'statusMessage': statusMessage,
-      'responseTime': responseTime?.toIso8601String(),
+      'responseTime': responseTime?.toUtc().toString(),
       'duration': duration,
       'headers': changedHeaders,
       'data': changedData,
