@@ -2,6 +2,10 @@ import 'dart:io';
 
 import 'package:cr_logger/cr_logger.dart';
 import 'package:cr_logger/src/cr_logger_helper.dart';
+import 'package:cr_logger/src/data/sqflite_db/converters/http_enitity_converter.dart';
+import 'package:cr_logger/src/data/sqflite_db/converters/log_entity_converters.dart';
+import 'package:cr_logger/src/data/sqflite_db/entities/http_entity.dart';
+import 'package:cr_logger/src/data/sqflite_db/entities/log_entity.dart';
 import 'package:cr_logger/src/js/console_output_worker.dart';
 import 'package:cr_logger/src/js/scripts.dart';
 import 'package:cr_logger/src/managers/log_manager.dart';
@@ -18,6 +22,9 @@ class TransferManager {
 
   final _useDB = CRLoggerHelper.instance.useDB;
   final _parser = IsolateParser();
+
+  final _logCnv = LogEntityConverter();
+  final _httpCnv = HttpEntityConverter();
 
   Future<void> createJsonFileAndShare() async {
     final json = await _parser.encode(await _toJson());
@@ -48,79 +55,121 @@ class TransferManager {
     final json = await _parser.decode(await file.readAsString());
 
     await _logMng.clean();
-    _setLogsFromJson(json);
+    await _setLogsFromJson(json);
     _httpMng.update();
     _logMng.onAllUpdate?.call();
   }
 
+  /// Works only in Web.
   Future<void> createLogsFromJson(Map<String, dynamic> json) async {
     await _logMng.clean();
-    _setLogsFromJson(json);
+    await _setLogsFromJson(json);
     _httpMng.update();
     _logMng.onAllUpdate?.call();
+  }
+
+  Future<List<Map<String, dynamic>>> _httpLogsToJson(
+    List<HttpBean> logs,
+  ) async {
+    final httpLogs = <Map<String, dynamic>>[];
+    for (final log in logs) {
+      final httpEntity = await _httpCnv.outToIn(log);
+      httpLogs.add(httpEntity.toJson());
+    }
+
+    return httpLogs;
   }
 
   Future<Map<String, dynamic>> _toJson() async {
-    final httpLog = <String, dynamic>{};
-
     if (_useDB) {
       await _logMng.loadLogsFromDB(getWithCurrentLogs: true);
       await _httpMng.loadLogsFromDB(getWithCurrentLogs: true);
-      for (final element in _httpMng.logsFromDB) {
-        httpLog[element.key.toString()] = element.toJson();
-      }
-    } else {
-      for (final element in _httpMng.logMap.entries) {
-        httpLog[element.key] = element.value.toJson();
-      }
+      await _httpLogsToJson(_httpMng.logsFromDB);
     }
 
     return {
-      'http': httpLog,
-      'debug': _logModelsToJson(_useDB ? _logMng.logDebugDB : _logMng.logDebug),
-      'info': _logModelsToJson(_useDB ? _logMng.logInfoDB : _logMng.logInfo),
-      'error': _logModelsToJson(_useDB ? _logMng.logErrorDB : _logMng.logError),
+      'http': await _httpLogsToJson(
+        _useDB
+            ? _httpMng.logsFromDB
+            : _httpMng.logMap.entries.map((e) => e.value).toList(),
+      ),
+      'debug': await _logModelsToJson(
+        _useDB ? _logMng.logDebugDB : _logMng.logDebug,
+      ),
+      'info': await _logModelsToJson(
+        _useDB ? _logMng.logInfoDB : _logMng.logInfo,
+      ),
+      'error': await _logModelsToJson(
+        _useDB ? _logMng.logErrorDB : _logMng.logError,
+      ),
     };
   }
 
-  void _setLogsFromJson(Map<String, dynamic> json) {
-    if (json['http'] != null) {
-      final httpLogs = json['http'] as Map<String, dynamic>;
-      for (final element in httpLogs.entries) {
-        _httpMng.logMap[element.key] = HttpBean.fromJson(element.value);
-        _httpMng.keys.add(element.key);
-      }
-    }
+  Future<void> _setLogsFromJson(Map<String, dynamic> json) async {
+    final http = json['http'];
     final debug = json['debug'];
     final info = json['info'];
     final error = json['error'];
+
+    if (http is List) {
+      await _httpLogsFromJson(
+        http.map((e) => e as Map<String, dynamic>).toList(),
+      );
+    }
     if (debug is List) {
-      _logMng.logDebug.addAll(_logModelsFromJson(
-        debug.map((e) => e as Map<String, dynamic>).toList(),
-      ));
+      _logMng.logDebug.addAll(
+        await _logModelsFromJson(
+          debug.map((e) => e as Map<String, dynamic>).toList(),
+        ),
+      );
     }
     if (info is List) {
-      _logMng.logInfo.addAll(_logModelsFromJson(
-        info.map((e) => e as Map<String, dynamic>).toList(),
-      ));
+      _logMng.logInfo.addAll(
+        await _logModelsFromJson(
+          info.map((e) => e as Map<String, dynamic>).toList(),
+        ),
+      );
     }
     if (error is List) {
-      _logMng.logError.addAll(_logModelsFromJson(
-        error.map((e) => e as Map<String, dynamic>).toList(),
-      ));
+      _logMng.logError.addAll(
+        await _logModelsFromJson(
+          error.map((e) => e as Map<String, dynamic>).toList(),
+        ),
+      );
     }
   }
 
-  List<Map<String, dynamic>> _logModelsToJson(List<LogBean> logs) =>
-      logs.map((item) => item.toJson()).toList().reversed.toList();
+  Future<List<Map<String, dynamic>>> _logModelsToJson(
+    List<LogBean> logs,
+  ) async {
+    final jsonData = <Map<String, dynamic>>[];
+    for (final log in logs) {
+      final logEntity = await _logCnv.outToIn(log);
+      jsonData.add(logEntity.toJson());
+    }
 
-  List<LogBean> _logModelsFromJson(List<Map<String, dynamic>> json) {
+    return jsonData.toList();
+  }
+
+  Future<List<LogBean>> _logModelsFromJson(
+    List<Map<String, dynamic>> json,
+  ) async {
     final listOfLogBean = <LogBean>[];
 
-    for (final logBean in json) {
-      listOfLogBean.add(LogBean.fromJson(logBean));
+    for (final data in json) {
+      final logBean = await _logCnv.inToOut(LogEntity.fromJson(data));
+      listOfLogBean.add(logBean);
     }
 
     return listOfLogBean;
+  }
+
+  Future<void> _httpLogsFromJson(List<Map<String, dynamic>> json) async {
+    for (final element in json) {
+      final httpBean = await _httpCnv.inToOut(HttpEntity.fromJson(element));
+      final key = httpBean.key.toString();
+      _httpMng.logMap[key] = httpBean;
+      _httpMng.keys.add(key);
+    }
   }
 }
