@@ -1,4 +1,5 @@
 import 'package:cr_logger/src/constants.dart';
+import 'package:cr_logger/src/controllers/logs_mode_controller.dart';
 import 'package:cr_logger/src/data/bean/http_bean.dart';
 import 'package:cr_logger/src/extensions/extensions.dart';
 import 'package:cr_logger/src/models/request_status.dart';
@@ -7,6 +8,7 @@ import 'package:cr_logger/src/res/styles.dart';
 import 'package:cr_logger/src/utils/parsers/url_parser.dart';
 import 'package:cr_logger/src/widget/copy_widget.dart';
 import 'package:cr_logger/src/widget/rounded_card.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -33,6 +35,14 @@ class HttpItem extends StatelessWidget {
     final urlWithHiddenParams =
         getUrlWithHiddenParams(httpBean.request?.url ?? '');
 
+    final request = httpBean.request;
+
+    /// [!kIsWeb] because the logs may be imported
+    /// and then there is no way to know the date of the logs.
+    final time = LogsModeController.instance.isFromCurrentSession && !kIsWeb
+        ? request?.requestTime?.formatTime()
+        : request?.requestTime?.formatTimeWithYear();
+
     return RoundedCard(
       padding: const EdgeInsets.only(
         left: 16,
@@ -49,7 +59,7 @@ class HttpItem extends StatelessWidget {
               /// Request method
               SizedBox(
                 child: Text(
-                  httpBean.request?.method ?? '',
+                  request?.method ?? '',
                   style:
                       CRStyle.bodyBlackSemiBold14.copyWith(color: methodColor),
                   textAlign: TextAlign.center,
@@ -106,7 +116,7 @@ class HttpItem extends StatelessWidget {
 
           /// Time and duration
           Text(
-            '${httpBean.request?.requestTime?.formatTime()}  duration: ${httpBean.response?.duration ?? httpBean.error?.duration ?? 0}ms',
+            '$time duration: ${httpBean.response?.duration ?? httpBean.error?.duration ?? 0}ms',
             style: CRStyle.bodyGreyRegular14,
           ),
         ],
